requirements.txt
Django==3.0.2
djangorestframework==3.12.2
django-filter==2.4.0
celery==5.2.0